requirements.txt
py-clob-client>=0.20.0
pandas>=2.0
numpy>=1.24
websockets>=12.0
aiohttp>=3.9
python-dotenv>=1.0
requests>=2.31
web3>=6.0
pyyaml>=6.0
